# swagger_rcs_gateway.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RCS Gateway API with Authentication",
...
        "contact": {
            "name": "Suporte RCS Gateway",
            "email": "suporte@exemplo.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/rcs/basic": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Send Basic Message Legacy",
                "description": "[LEGACY] Envio de mensagem RCS Basic (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/send-message` para novos projetos",
                "operationId": "send_basic_message_legacy_api_rcs_basic_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RCSBasicRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rcs/single": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Send Single Message Legacy",
                "description": "[LEGACY] Envio de mensagem RCS Single (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/send-message` para novos projetos",
                "operationId": "send_single_message_legacy_api_rcs_single_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RCSSingleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages": {
            "get": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "List Messages Legacy",
                "description": "[LEGACY] Listar mensagens (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/messages` para novos projetos",
                "operationId": "list_messages_legacy_api_messages_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rcs/callback": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Receive Callback Legacy",
                "description": "[LEGACY] Receber callback da API RCS (mantido para compatibilidade)\n\nEste endpoint ainda \u00e9 usado pela Eugen para enviar callbacks.",
                "operationId": "receive_callback_legacy_api_rcs_callback_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "title": "Callback Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Fallback": {
                "properties": {
                    "channel": {
                        "type": "string",
                        "title": "Channel",
                        "default": "SMS"
                    },
                    "content": {
                        "type": "string",
                        "maxLength": 160,
                        "title": "Content",
                        "description": "Mensagem SMS sem acentos"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "Fallback"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ImageContent": {
                "properties": {
                    "image": {
                        "type": "string",
                        "title": "Image",
                        "description": "URL HTTPS da imagem"
                    }
                },
                "type": "object",
                "required": [
                    "image"
                ],
                "title": "ImageContent"
            },
...
            "MessageContent": {
                "properties": {
                    "text": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/TextContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "image": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ImageContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "video": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/VideoContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "pdf": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PDFContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "suggestion": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/SuggestionContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "richCard": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/RichCardContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "carousel": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/CarouselItem"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Carousel"
                    }
                },
                "type": "object",
                "title": "MessageContent"
            },
            "MessageRecipient": {
                "properties": {
                    "number": {
                        "type": "string",
                        "title": "Number",
                        "description": "N\u00famero do telefone"
                    },
                    "vars": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "type": "string"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Vars",
                        "description": "Vari\u00e1veis din\u00e2micas"
                    }
                },
                "type": "object",
                "required": [
                    "number"
                ],
                "title": "MessageRecipient"
            },
...
            "RCSBasicRequest": {
                "properties": {
                    "campaign_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Campaign Name"
                    },
                    "account": {
                        "type": "string",
                        "title": "Account"
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/MessageRecipient"
                        },
                        "type": "array",
                        "title": "Messages"
                    },
                    "content": {
                        "$ref": "#/components/schemas/MessageContent"
                    },
                    "callback": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Callback"
                    },
                    "fallback": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Fallback"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fallback"
                    }
                },
                "type": "object",
                "required": [
                    "account",
                    "messages",
                    "content"
                ],
                "title": "RCSBasicRequest"
            },
            "RCSSingleRequest": {
                "properties": {
                    "campaign_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Campaign Name"
                    },
                    "account": {
                        "type": "string",
                        "title": "Account"
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/MessageRecipient"
                        },
                        "type": "array",
                        "title": "Messages"
                    },
                    "template_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Template Id"
                    },
                    "content": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/MessageContent"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "callback": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Callback"
                    },
                    "fallback": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Fallback"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fallback"
                    }
                },
                "type": "object",
                "required": [
                    "account",
                    "messages"
                ],
                "title": "RCSSingleRequest"
            },
...
            "TextContent": {
                "properties": {
                    "message": {
                        "type": "string",
                        "maxLength": 5000,
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "TextContent"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}